+++ src/lib.rs
@@ -12,10 +12,21 @@ use core::fmt;
 
 mod common;
 mod signed;
+pub mod traits;
 mod unsigned;
+mod v1_number_compat;
 
 pub use signed::*;
 pub use unsigned::*;
+pub use v1_number_compat::*;
+
+/// The preferred way to import arbitrary-int into a project: `use arbitrary_int::prelude::*`
+pub mod prelude {
+    pub use crate::signed::*;
+    pub use crate::traits::*;
+    pub use crate::unsigned::*;
+    pub use crate::TryNewError;
+}
 
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub struct TryNewError;