+++ src/signed.rs
@@ -1,91 +1,28 @@
 use crate::{
-    common::{bytes_operation_impl, from_arbitrary_int_impl, from_native_impl, impl_extract},
+    traits::{Integer, SignedInteger, sealed::Sealed},
+    common::{
+        bytes_operation_impl, from_arbitrary_int_impl, from_native_impl, impl_extract,
+    },
     TryNewError,
 };
 use core::{
-    fmt::{self, Debug},
+    fmt::{self},
     ops::{
         Add, AddAssign, BitAnd, BitAndAssign, BitOr, BitOrAssign, BitXor, BitXorAssign, Div,
         DivAssign, Mul, MulAssign, Neg, Not, Shl, ShlAssign, Shr, ShrAssign, Sub, SubAssign,
     },
 };
 
-#[cfg_attr(feature = "const_convert_and_const_trait_impl", const_trait)]
-pub trait SignedNumber: Sized + Copy + Clone + PartialOrd + Ord + PartialEq + Eq {
-    type UnderlyingType: SignedNumber
-        + Debug
-        + From<i8>
-        + TryFrom<i16>
-        + TryFrom<i32>
-        + TryFrom<i64>
-        + TryFrom<i128>;
-
-    /// Number of bits that can fit in this type
-    const BITS: usize;
-
-    /// Minimum value that can be represented by this type
-    const MIN: Self;
-
-    /// Maximum value that can be represented by this type
-    const MAX: Self;
-
-    /// Creates a number from the given value, throwing an error if the value is too large.
-    /// This constructor is useful when creating a value from a literal.
-    fn new(value: Self::UnderlyingType) -> Self;
-
-    /// Creates a number from the given value, return None if the value is too large
-    fn try_new(value: Self::UnderlyingType) -> Result<Self, TryNewError>;
-
-    /// Returns the type as a fundamental data type.
-    ///
-    /// Note that if negative, the returned value may span more bits than [`BITS`](Self::BITS),
-    /// as it preserves the numeric value instead of the bitwise value:
-    ///
-    /// ```
-    /// # use arbitrary_int::i3;
-    /// let value: i8 = i3::new(-1).value();
-    /// assert_eq!(value, -1);
-    /// assert_eq!(value.count_ones(), 8);
-    /// ```
-    ///
-    /// If you need a value within the specified bit range, use [`Int::to_bits`].
-    fn value(self) -> Self::UnderlyingType;
-
-    /// Creates a number from the given value, throwing an error if the value is too large.
-    /// This constructor is useful when the value is convertible to T. Use [`Self::new`] for literals.
-    #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-    fn from_<T: SignedNumber>(value: T) -> Self;
-
-    /// Creates an instance from the given `value`. Unlike the various `new...` functions, this
-    /// will never fail as the value is masked to the result size.
-    #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-    fn masked_new<T: SignedNumber>(value: T) -> Self;
-
-    fn as_i8(&self) -> i8;
-
-    fn as_i16(&self) -> i16;
-
-    fn as_i32(&self) -> i32;
-
-    fn as_i64(&self) -> i64;
-
-    fn as_i128(&self) -> i128;
-
-    fn as_isize(&self) -> isize;
-
-    #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-    #[inline]
-    fn as_<T: SignedNumber>(self) -> T {
-        T::masked_new(self)
-    }
-}
-
-macro_rules! impl_signed_number_native {
+macro_rules! impl_signed_integer_native {
     // `$const_keyword` is marked as an optional fragment here so that it can conditionally be put on the impl.
     // This macro will be invoked with `i8 as const, ...` if `const_convert_and_const_trait_impl` is enabled.
     ($($type:ident $(as $const_keyword:ident)?),+) => {
         $(
-            impl $($const_keyword)? SignedNumber for $type {
+            impl $($const_keyword)? Sealed for $type {}
+
+            impl $($const_keyword)? SignedInteger for $type {}
+
+            impl $($const_keyword)? Integer for $type {
                 type UnderlyingType = $type;
                 const BITS: usize = Self::BITS as usize;
                 const MIN: Self = Self::MIN;
@@ -102,7 +39,7 @@ macro_rules! impl_signed_number_native {
 
                 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
                 #[inline]
-                fn from_<T: SignedNumber>(value: T) -> Self {
+                fn from_<T: Integer>(value: T) -> Self {
                     if T::BITS > Self::BITS as usize {
                         assert!(value >= T::masked_new(Self::MIN) && value <= T::masked_new(Self::MAX));
                     }
@@ -111,7 +48,7 @@ macro_rules! impl_signed_number_native {
 
                 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
                 #[inline]
-                fn masked_new<T: SignedNumber>(value: T) -> Self {
+                fn masked_new<T: Integer>(value: T) -> Self {
                     // Primitive types don't need masking
                     match Self::BITS {
                         8 => value.as_i8() as Self,
@@ -119,37 +56,55 @@ macro_rules! impl_signed_number_native {
                         32 => value.as_i32() as Self,
                         64 => value.as_i64() as Self,
                         128 => value.as_i128() as Self,
-                        _ => panic!("Unhandled Number type")
+                        _ => panic!("Unhandled Integer type")
                     }
                 }
 
                 #[inline]
-                fn as_i8(&self) -> i8 { *self as i8 }
+                fn as_u8(self) -> u8 { self as u8 }
 
                 #[inline]
-                fn as_i16(&self) -> i16 { *self as i16 }
+                fn as_u16(self) -> u16 { self as u16 }
 
                 #[inline]
-                fn as_i32(&self) -> i32 { *self as i32 }
+                fn as_u32(self) -> u32 { self as u32 }
 
                 #[inline]
-                fn as_i64(&self) -> i64 { *self as i64 }
+                fn as_u64(self) -> u64 { self as u64 }
 
                 #[inline]
-                fn as_i128(&self) -> i128 { *self as i128 }
+                fn as_u128(self) -> u128 { self as u128 }
 
                 #[inline]
-                fn as_isize(&self) -> isize { *self as isize }
+                fn as_usize(self) -> usize { self as usize }
+
+                #[inline]
+                fn as_i8(self) -> i8 { self as i8 }
+
+                #[inline]
+                fn as_i16(self) -> i16 { self as i16 }
+
+                #[inline]
+                fn as_i32(self) -> i32 { self as i32 }
+
+                #[inline]
+                fn as_i64(self) -> i64 { self as i64 }
+
+                #[inline]
+                fn as_i128(self) -> i128 { self as i128 }
+
+                #[inline]
+                fn as_isize(self) -> isize { self as isize }
             }
         )+
     };
 }
 
 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-impl_signed_number_native!(i8, i16, i32, i64, i128);
+impl_signed_integer_native!(i8, i16, i32, i64, i128);
 
 #[cfg(feature = "const_convert_and_const_trait_impl")]
-impl_signed_number_native!(i8 as const, i16 as const, i32 as const, i64 as const, i128 as const);
+impl_signed_integer_native!(i8 as const, i16 as const, i32 as const, i64 as const, i128 as const);
 
 #[derive(Copy, Clone, Eq, PartialEq, Default, Ord, PartialOrd)]
 pub struct Int<T, const BITS: usize> {
@@ -197,7 +152,11 @@ macro_rules! int_impl_num {
     // This macro will be invoked with `i8 as const, ...` if `const_convert_and_const_trait_impl` is enabled.
     ($($type:ident $(as $const_keyword:ident)?),+) => {
         $(
-            impl<const BITS: usize> $($const_keyword)? SignedNumber for Int<$type, BITS> {
+            impl<const BITS: usize> $($const_keyword)? Sealed for Int<$type, BITS> {}
+
+            impl<const BITS: usize> $($const_keyword)? SignedInteger for Int<$type, BITS> {}
+
+            impl<const BITS: usize> $($const_keyword)? Integer for Int<$type, BITS> {
                 type UnderlyingType = $type;
 
                 const BITS: usize = BITS;
@@ -208,7 +167,7 @@ macro_rules! int_impl_num {
                 // we will get a compiler error right here
                 const MAX: Self = Self {
                     // MAX is always positive so we don't have to worry about the sign
-                    value: (<$type as SignedNumber>::MAX >> (<$type as SignedNumber>::BITS - Self::BITS)),
+                    value: (<$type as Integer>::MAX >> (<$type as Integer>::BITS - Self::BITS)),
                 };
 
                 #[inline]
@@ -229,7 +188,7 @@ macro_rules! int_impl_num {
 
                 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
                 #[inline]
-                fn from_<T: SignedNumber>(value: T) -> Self {
+                fn from_<T: Integer>(value: T) -> Self {
                     if Self::BITS < T::BITS {
                         assert!(value >= Self::MIN.value.as_() && value <= Self::MAX.value.as_());
                     }
@@ -237,7 +196,7 @@ macro_rules! int_impl_num {
                 }
 
                 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-                fn masked_new<T: SignedNumber>(value: T) -> Self {
+                fn masked_new<T: Integer>(value: T) -> Self {
                     if Self::BITS < T::BITS {
                         let value = (value.as_::<Self::UnderlyingType>() << Self::UNUSED_BITS) >> Self::UNUSED_BITS;
                         Self { value: Self::UnderlyingType::masked_new(value) }
@@ -246,27 +205,51 @@ macro_rules! int_impl_num {
                     }
                 }
 
-                fn as_i8(&self) -> i8 {
+                fn as_u8(self) -> u8 {
                     self.value() as _
                 }
 
-                fn as_i16(&self) -> i16 {
+                fn as_u16(self) -> u16 {
                     self.value() as _
                 }
 
-                fn as_i32(&self) -> i32 {
+                fn as_u32(self) -> u32 {
                     self.value() as _
                 }
 
-                fn as_i64(&self) -> i64 {
+                fn as_u64(self) -> u64 {
                     self.value() as _
                 }
 
-                fn as_i128(&self) -> i128 {
+                fn as_u128(self) -> u128 {
                     self.value() as _
                 }
 
-                fn as_isize(&self) -> isize {
+                fn as_usize(self) -> usize {
+                    self.value() as _
+                }
+
+                fn as_i8(self) -> i8 {
+                    self.value() as _
+                }
+
+                fn as_i16(self) -> i16 {
+                    self.value() as _
+                }
+
+                fn as_i32(self) -> i32 {
+                    self.value() as _
+                }
+
+                fn as_i64(self) -> i64 {
+                    self.value() as _
+                }
+
+                fn as_i128(self) -> i128 {
+                    self.value() as _
+                }
+
+                fn as_isize(self) -> isize {
                     self.value() as _
                 }
 
@@ -500,7 +483,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(100).wrapping_add(i14::new(27)), i14::new(127));
                 /// assert_eq!(i14::MAX.wrapping_add(i14::new(2)), i14::MIN + i14::new(1));
                 /// ```
@@ -521,7 +504,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(0).wrapping_sub(i14::new(127)), i14::new(-127));
                 /// assert_eq!(i14::new(-2).wrapping_sub(i14::MAX), i14::MAX);
                 /// ```
@@ -542,7 +525,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i14;
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(10).wrapping_mul(i14::new(12)), i14::new(120));
                 /// assert_eq!(i14::new(12).wrapping_mul(i14::new(1024)), i14::new(-4096));
                 /// ```
@@ -572,7 +555,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(100).wrapping_div(i14::new(10)), i14::new(10));
                 /// assert_eq!(i14::MIN.wrapping_div(i14::new(-1)), i14::MIN);
                 /// ```
@@ -598,7 +581,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(100).wrapping_neg(), i14::new(-100));
                 /// assert_eq!(i14::new(-100).wrapping_neg(), i14::new(100));
                 /// assert_eq!(i14::MIN.wrapping_neg(), i14::MIN);
@@ -624,7 +607,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i14;
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(-1).wrapping_shl(7), i14::new(-128));
                 /// assert_eq!(i14::new(-1).wrapping_shl(128), i14::new(-4));
                 /// ```
@@ -661,7 +644,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i14;
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(-128).wrapping_shr(7), i14::new(-1));
                 /// assert_eq!(i14::new(-128).wrapping_shr(60), i14::new(-8));
                 /// ```
@@ -688,7 +671,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(100).saturating_add(i14::new(1)), i14::new(101));
                 /// assert_eq!(i14::MAX.saturating_add(i14::new(100)), i14::MAX);
                 /// assert_eq!(i14::MIN.saturating_add(i14::new(-1)), i14::MIN);
@@ -724,7 +707,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(100).saturating_sub(i14::new(127)), i14::new(-27));
                 /// assert_eq!(i14::MIN.saturating_sub(i14::new(100)), i14::MIN);
                 /// assert_eq!(i14::MAX.saturating_sub(i14::new(-1)), i14::MAX);
@@ -760,7 +743,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(10).saturating_mul(i14::new(12)), i14::new(120));
                 /// assert_eq!(i14::MAX.saturating_mul(i14::new(10)), i14::MAX);
                 /// assert_eq!(i14::MIN.saturating_mul(i14::new(10)), i14::MIN);
@@ -798,7 +781,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(5).saturating_div(i14::new(2)), i14::new(2));
                 /// assert_eq!(i14::MAX.saturating_div(i14::new(-1)), i14::MIN + i14::new(1));
                 /// assert_eq!(i14::MIN.saturating_div(i14::new(-1)), i14::MAX);
@@ -826,7 +809,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(100).saturating_neg(), i14::new(-100));
                 /// assert_eq!(i14::new(-100).saturating_neg(), i14::new(100));
                 /// assert_eq!(i14::MIN.saturating_neg(), i14::MAX);
@@ -852,7 +835,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(-4).saturating_pow(3), i14::new(-64));
                 /// assert_eq!(i14::MIN.saturating_pow(2), i14::MAX);
                 /// assert_eq!(i14::MIN.saturating_pow(3), i14::MIN);
@@ -880,7 +863,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!((i14::MAX - i14::new(2)).checked_add(i14::new(1)), Some(i14::MAX - i14::new(1)));
                 /// assert_eq!((i14::MAX - i14::new(2)).checked_add(i14::new(3)), None);
                 /// ```
@@ -914,7 +897,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!((i14::MIN + i14::new(2)).checked_sub(i14::new(1)), Some(i14::MIN + i14::new(1)));
                 /// assert_eq!((i14::MIN + i14::new(2)).checked_sub(i14::new(3)), None);
                 /// ```
@@ -948,7 +931,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::MAX.checked_mul(i14::new(1)), Some(i14::MAX));
                 /// assert_eq!(i14::MAX.checked_mul(i14::new(2)), None);
                 /// ```
@@ -980,7 +963,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!((i14::MIN + i14::new(1)).checked_div(i14::new(-1)), Some(i14::new(8191)));
                 /// assert_eq!(i14::MIN.checked_div(i14::new(-1)), None);
                 /// assert_eq!((i14::new(1)).checked_div(i14::new(0)), None);
@@ -1004,7 +987,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i14, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i14::new(5).checked_neg(), Some(i14::new(-5)));
                 /// assert_eq!(i14::MIN.checked_neg(), None);
                 /// ```
@@ -1107,7 +1090,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i6;
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i6::from_bits(0b10_1010).reverse_bits(), i6::from_bits(0b01_0101));
                 /// assert_eq!(i6::new(0), i6::new(0).reverse_bits());
                 /// ```
@@ -1125,7 +1108,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i6;
+                /// # use arbitrary_int::prelude::*;
                 /// let n = i6::from_bits(0b00_1000);
                 /// assert_eq!(n.count_ones(), 1);
                 /// ```
@@ -1142,7 +1125,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{i6, SignedNumber};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(i6::MAX.count_zeros(), 1);
                 /// ```
                 #[inline]
@@ -1160,7 +1143,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i6;
+                /// # use arbitrary_int::prelude::*;
                 /// let n = i6::new(-1);
                 /// assert_eq!(n.leading_ones(), 6);
                 /// ```
@@ -1176,7 +1159,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i6;
+                /// # use arbitrary_int::prelude::*;
                 /// let n = i6::new(-1);
                 /// assert_eq!(n.leading_zeros(), 0);
                 /// ```
@@ -1199,7 +1182,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i6;
+                /// # use arbitrary_int::prelude::*;
                 /// let n = i6::new(3);
                 /// assert_eq!(n.trailing_ones(), 2);
                 /// ```
@@ -1217,7 +1200,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i6;
+                /// # use arbitrary_int::prelude::*;
                 /// let n = i6::new(-4);
                 /// assert_eq!(n.trailing_zeros(), 2);
                 /// ```
@@ -1238,7 +1221,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i6;
+                /// # use arbitrary_int::prelude::*;
                 /// let n = i6::from_bits(0b10_1010);
                 /// let m = i6::from_bits(0b01_0101);
                 ///
@@ -1267,7 +1250,7 @@ macro_rules! int_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::i6;
+                /// # use arbitrary_int::prelude::*;
                 /// let n = i6::from_bits(0b10_1010);
                 /// let m = i6::from_bits(0b01_0101);
                 ///
@@ -1305,7 +1288,7 @@ int_impl!(
 // Arithmetic operator implementations
 impl<T, const BITS: usize> Add for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + Add<T, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
 {
     type Output = Self;
@@ -1320,7 +1303,7 @@ where
 
 impl<T, const BITS: usize> AddAssign for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + Add<T, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
 {
     fn add_assign(&mut self, rhs: Self) {
@@ -1331,7 +1314,7 @@ where
 
 impl<T, const BITS: usize> Sub for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + Sub<T, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
 {
     type Output = Self;
@@ -1346,7 +1329,7 @@ where
 
 impl<T, const BITS: usize> SubAssign for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + Sub<T, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
 {
     fn sub_assign(&mut self, rhs: Self) {
@@ -1357,7 +1340,7 @@ where
 
 impl<T, const BITS: usize> Mul for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + Mul<T, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
 {
     type Output = Self;
@@ -1372,7 +1355,7 @@ where
 
 impl<T, const BITS: usize> MulAssign for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + Mul<T, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
 {
     fn mul_assign(&mut self, rhs: Self) {
@@ -1383,7 +1366,7 @@ where
 
 impl<T, const BITS: usize> Div for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + Div<T, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
 {
     type Output = Self;
@@ -1400,7 +1383,7 @@ where
 
 impl<T, const BITS: usize> DivAssign for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + Div<T, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
 {
     fn div_assign(&mut self, rhs: Self) {
@@ -1411,7 +1394,7 @@ where
 
 impl<T, const BITS: usize> Neg for Int<T, BITS>
 where
-    Self: SignedNumber<UnderlyingType = T>,
+    Self: Integer<UnderlyingType = T>,
     T: PartialEq + Copy + Neg<Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
 {
     type Output = Self;
@@ -1428,7 +1411,7 @@ where
 // Bitwise operator implementations
 impl<T, const BITS: usize> BitAnd for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + BitAnd<T, Output = T>,
 {
     type Output = Self;
@@ -1441,7 +1424,7 @@ where
 
 impl<T, const BITS: usize> BitAndAssign for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + BitAndAssign<T>,
 {
     fn bitand_assign(&mut self, rhs: Self) {
@@ -1451,7 +1434,7 @@ where
 
 impl<T, const BITS: usize> BitOr for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + BitOr<T, Output = T>,
 {
     type Output = Self;
@@ -1464,7 +1447,7 @@ where
 
 impl<T, const BITS: usize> BitOrAssign for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + BitOrAssign<T>,
 {
     fn bitor_assign(&mut self, rhs: Self) {
@@ -1474,7 +1457,7 @@ where
 
 impl<T, const BITS: usize> BitXor for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + BitXor<T, Output = T>,
 {
     type Output = Self;
@@ -1487,7 +1470,7 @@ where
 
 impl<T, const BITS: usize> BitXorAssign for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + BitXorAssign<T>,
 {
     fn bitxor_assign(&mut self, rhs: Self) {
@@ -1497,7 +1480,7 @@ where
 
 impl<T, const BITS: usize> Not for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: PartialEq + Copy + Not<Output = T>,
 {
     type Output = Self;
@@ -1510,7 +1493,7 @@ where
 
 impl<T, TSHIFTBITS, const BITS: usize> Shl<TSHIFTBITS> for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: Copy + Shl<TSHIFTBITS, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
     TSHIFTBITS: TryInto<usize> + Copy,
 {
@@ -1533,7 +1516,7 @@ where
 
 impl<T, TSHIFTBITS, const BITS: usize> ShlAssign<TSHIFTBITS> for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: Copy + Shl<TSHIFTBITS, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
     TSHIFTBITS: TryInto<usize> + Copy,
 {
@@ -1545,7 +1528,7 @@ where
 
 impl<T, TSHIFTBITS, const BITS: usize> Shr<TSHIFTBITS> for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: Copy + Shr<TSHIFTBITS, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
     TSHIFTBITS: TryInto<usize> + Copy,
 {
@@ -1569,7 +1552,7 @@ where
 
 impl<T, TSHIFTBITS, const BITS: usize> ShrAssign<TSHIFTBITS> for Int<T, BITS>
 where
-    Self: SignedNumber,
+    Self: Integer,
     T: Copy + Shr<TSHIFTBITS, Output = T> + Shl<usize, Output = T> + Shr<usize, Output = T>,
     TSHIFTBITS: TryInto<usize> + Copy,
 {
@@ -1659,7 +1642,7 @@ where
 #[cfg(feature = "serde")]
 struct InvalidIntValueError<T>
 where
-    T: SignedNumber,
+    T: Integer,
     T::UnderlyingType: fmt::Display,
 {
     value: T::UnderlyingType,
@@ -1668,7 +1651,7 @@ where
 #[cfg(feature = "serde")]
 impl<T> fmt::Display for InvalidIntValueError<T>
 where
-    T: SignedNumber,
+    T: Integer,
     T::UnderlyingType: fmt::Display,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
@@ -1695,7 +1678,7 @@ where
 #[cfg(feature = "serde")]
 impl<'de, T, const BITS: usize> serde::Deserialize<'de> for Int<T, BITS>
 where
-    Self: SignedNumber<UnderlyingType = T>,
+    Self: Integer<UnderlyingType = T>,
     T: fmt::Display + PartialOrd + serde::Deserialize<'de>,
 {
     fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {