+++ src/unsigned.rs
@@ -1,6 +1,7 @@
 use crate::common::{
     bytes_operation_impl, from_arbitrary_int_impl, from_native_impl, impl_extract,
 };
+use crate::traits::{Integer, UnsignedInteger, sealed::Sealed};
 use crate::TryNewError;
 use core::fmt::{Binary, Debug, Display, Formatter, LowerHex, Octal, UpperHex};
 #[cfg(feature = "step_trait")]
@@ -24,69 +25,21 @@ use std::{collections::BTreeMap, string::ToString};
 #[cfg(feature = "schemars")]
 use schemars::JsonSchema;
 
-#[cfg_attr(feature = "const_convert_and_const_trait_impl", const_trait)]
-pub trait Number: Sized + Copy + Clone + PartialOrd + Ord + PartialEq + Eq {
-    type UnderlyingType: Number
-        + Debug
-        + From<u8>
-        + TryFrom<u16>
-        + TryFrom<u32>
-        + TryFrom<u64>
-        + TryFrom<u128>;
-
-    /// Number of bits that can fit in this type
-    const BITS: usize;
-
-    /// Minimum value that can be represented by this type
-    const MIN: Self;
-
-    /// Maximum value that can be represented by this type
-    const MAX: Self;
-
-    /// Creates a number from the given value, throwing an error if the value is too large.
-    /// This constructor is useful when creating a value from a literal.
-    fn new(value: Self::UnderlyingType) -> Self;
-
-    /// Creates a number from the given value, return None if the value is too large
-    fn try_new(value: Self::UnderlyingType) -> Result<Self, TryNewError>;
-
-    fn value(self) -> Self::UnderlyingType;
-
-    /// Creates a number from the given value, throwing an error if the value is too large.
-    /// This constructor is useful when the value is convertible to T. Use [`Self::new`] for literals.
-    #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-    fn from_<T: Number>(value: T) -> Self;
-
-    /// Creates an instance from the given `value`. Unlike the various `new...` functions, this
-    /// will never fail as the value is masked to the result size.
-    #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-    fn masked_new<T: Number>(value: T) -> Self;
-
-    fn as_u8(&self) -> u8;
-
-    fn as_u16(&self) -> u16;
-
-    fn as_u32(&self) -> u32;
-
-    fn as_u64(&self) -> u64;
-
-    fn as_u128(&self) -> u128;
-
-    fn as_usize(&self) -> usize;
-
-    #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-    #[inline]
-    fn as_<T: Number>(self) -> T {
-        T::masked_new(self)
-    }
-}
-
-macro_rules! impl_number_native {
+macro_rules! impl_integer_native {
     // `$const_keyword` is marked as an optional fragment here so that it can conditionally be put on the impl.
     // This macro will be invoked with `u8 as const, ...` if `const_convert_and_const_trait_impl` is enabled.
     ($($type:ident $(as $const_keyword:ident)?),+) => {
         $(
-            impl $($const_keyword)? Number for $type {
+            #[allow(deprecated)]
+            impl crate::v1_number_compat::Number for $type {
+                type UnderlyingType = $type;
+            }
+
+            impl $($const_keyword)? Sealed for $type {}
+
+            impl $($const_keyword)? UnsignedInteger for $type {}
+
+            impl $($const_keyword)? Integer for $type {
                 type UnderlyingType = $type;
                 const BITS: usize = Self::BITS as usize;
                 const MIN: Self = Self::MIN;
@@ -103,7 +56,7 @@ macro_rules! impl_number_native {
 
                 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
                 #[inline]
-                fn from_<T: Number>(value: T) -> Self {
+                fn from_<T: Integer>(value: T) -> Self {
                     if T::BITS > Self::BITS as usize {
                         assert!(value <= T::masked_new(Self::MAX));
                     }
@@ -112,7 +65,7 @@ macro_rules! impl_number_native {
 
                 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
                 #[inline]
-                fn masked_new<T: Number>(value: T) -> Self {
+                fn masked_new<T: Integer>(value: T) -> Self {
                     // Primitive types don't need masking
                     match Self::BITS {
                         8 => value.as_u8() as Self,
@@ -120,37 +73,55 @@ macro_rules! impl_number_native {
                         32 => value.as_u32() as Self,
                         64 => value.as_u64() as Self,
                         128 => value.as_u128() as Self,
-                        _ => panic!("Unhandled Number type")
+                        _ => panic!("Unhandled Integer type")
                     }
                 }
 
                 #[inline]
-                fn as_u8(&self) -> u8 { *self as u8 }
+                fn as_u8(self) -> u8 { self as u8 }
 
                 #[inline]
-                fn as_u16(&self) -> u16 { *self as u16 }
+                fn as_u16(self) -> u16 { self as u16 }
 
                 #[inline]
-                fn as_u32(&self) -> u32 { *self as u32 }
+                fn as_u32(self) -> u32 { self as u32 }
 
                 #[inline]
-                fn as_u64(&self) -> u64 { *self as u64 }
+                fn as_u64(self) -> u64 { self as u64 }
 
                 #[inline]
-                fn as_u128(&self) -> u128 { *self as u128 }
+                fn as_u128(self) -> u128 { self as u128 }
 
                 #[inline]
-                fn as_usize(&self) -> usize { *self as usize }
+                fn as_usize(self) -> usize { self as usize }
+
+                #[inline]
+                fn as_i8(self) -> i8 { self as i8 }
+
+                #[inline]
+                fn as_i16(self) -> i16 { self as i16 }
+
+                #[inline]
+                fn as_i32(self) -> i32 { self as i32 }
+
+                #[inline]
+                fn as_i64(self) -> i64 { self as i64 }
+
+                #[inline]
+                fn as_i128(self) -> i128 { self as i128 }
+
+                #[inline]
+                fn as_isize(self) -> isize { self as isize }
             }
         )+
     };
 }
 
 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-impl_number_native!(u8, u16, u32, u64, u128);
+impl_integer_native!(u8, u16, u32, u64, u128);
 
 #[cfg(feature = "const_convert_and_const_trait_impl")]
-impl_number_native!(u8 as const, u16 as const, u32 as const, u64 as const, u128 as const);
+impl_integer_native!(u8 as const, u16 as const, u32 as const, u64 as const, u128 as const);
 
 #[derive(Copy, Clone, Eq, PartialEq, Default, Ord, PartialOrd, Hash)]
 pub struct UInt<T, const BITS: usize> {
@@ -182,7 +153,7 @@ impl<T: Copy, const BITS: usize> UInt<T, BITS> {
 
 impl<T, const BITS: usize> UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: Copy,
 {
     pub const MASK: T = Self::MAX.value;
@@ -198,7 +169,16 @@ macro_rules! uint_impl_num {
     // This macro will be invoked with `u8 as const, ...` if `const_convert_and_const_trait_impl` is enabled.
     ($($type:ident $(as $const_keyword:ident)?),+) => {
         $(
-            impl<const BITS: usize> $($const_keyword)? Number for UInt<$type, BITS> {
+            #[allow(deprecated)]
+            impl<const BITS: usize> crate::v1_number_compat::Number for UInt<$type, BITS> {
+                type UnderlyingType = $type;
+            }
+
+            impl<const BITS: usize> $($const_keyword)? Sealed for UInt<$type, BITS> {}
+
+            impl<const BITS: usize> $($const_keyword)? UnsignedInteger for UInt<$type, BITS> {}
+
+            impl<const BITS: usize> $($const_keyword)? Integer for UInt<$type, BITS> {
                 type UnderlyingType = $type;
 
                 const BITS: usize = BITS;
@@ -207,7 +187,7 @@ macro_rules! uint_impl_num {
 
                 // The existence of MAX also serves as a bounds check: If NUM_BITS is > available bits,
                 // we will get a compiler error right here
-                const MAX: Self = Self { value: (<$type as Number>::MAX >> (<$type as Number>::BITS - Self::BITS)) };
+                const MAX: Self = Self { value: (<$type as Integer>::MAX >> (<$type as Integer>::BITS - Self::BITS)) };
 
                 #[inline]
                 fn try_new(value: Self::UnderlyingType) -> Result<Self, TryNewError> {
@@ -227,7 +207,7 @@ macro_rules! uint_impl_num {
 
                 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
                 #[inline]
-                fn from_<T: Number>(value: T) -> Self {
+                fn from_<T: Integer>(value: T) -> Self {
                     if Self::BITS < T::BITS {
                         assert!(value <= Self::MAX.value.as_());
                     }
@@ -235,7 +215,7 @@ macro_rules! uint_impl_num {
                 }
 
                 #[cfg(not(feature = "const_convert_and_const_trait_impl"))]
-                fn masked_new<T: Number>(value: T) -> Self {
+                fn masked_new<T: Integer>(value: T) -> Self {
                     if Self::BITS < T::BITS {
                         Self { value: Self::UnderlyingType::masked_new(value.as_::<Self::UnderlyingType>() & Self::MASK) }
                     } else {
@@ -243,27 +223,51 @@ macro_rules! uint_impl_num {
                     }
                 }
 
-                fn as_u8(&self) -> u8 {
+                fn as_u8(self) -> u8 {
                     self.value() as _
                 }
 
-                fn as_u16(&self) -> u16 {
+                fn as_u16(self) -> u16 {
                     self.value() as _
                 }
 
-                fn as_u32(&self) -> u32 {
+                fn as_u32(self) -> u32 {
                     self.value() as _
                 }
 
-                fn as_u64(&self) -> u64 {
+                fn as_u64(self) -> u64 {
                     self.value() as _
                 }
 
-                fn as_u128(&self) -> u128 {
+                fn as_u128(self) -> u128 {
                     self.value() as _
                 }
 
-                fn as_usize(&self) -> usize {
+                fn as_usize(self) -> usize {
+                    self.value() as _
+                }
+
+                fn as_i8(self) -> i8 {
+                    self.value() as _
+                }
+
+                fn as_i16(self) -> i16 {
+                    self.value() as _
+                }
+
+                fn as_i32(self) -> i32 {
+                    self.value() as _
+                }
+
+                fn as_i64(self) -> i64 {
+                    self.value() as _
+                }
+
+                fn as_i128(self) -> i128 {
+                    self.value() as _
+                }
+
+                fn as_isize(self) -> isize {
                     self.value() as _
                 }
 
@@ -416,7 +420,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u14, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(u14::new(200).wrapping_add(u14::new(55)), u14::new(255));
                 /// assert_eq!(u14::new(200).wrapping_add(u14::MAX), u14::new(199));
                 /// ```
@@ -437,7 +441,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u14, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(u14::new(100).wrapping_sub(u14::new(100)), u14::new(0));
                 /// assert_eq!(u14::new(100).wrapping_sub(u14::MAX), u14::new(101));
                 /// ```
@@ -578,7 +582,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u14, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(u14::new(100).saturating_add(u14::new(1)), u14::new(101));
                 /// assert_eq!(u14::MAX.saturating_add(u14::new(100)), u14::MAX);
                 /// ```
@@ -632,7 +636,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u14, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(u14::new(2).saturating_mul(u14::new(10)), u14::new(20));
                 /// assert_eq!(u14::MAX.saturating_mul(u14::new(10)), u14::MAX);
                 /// ```
@@ -689,7 +693,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u14, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(u14::new(4).saturating_pow(3), u14::new(64));
                 /// assert_eq!(u14::MAX.saturating_pow(2), u14::MAX);
                 /// ```
@@ -713,7 +717,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u14, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!((u14::MAX - u14::new(2)).checked_add(u14::new(1)), Some(u14::MAX - u14::new(1)));
                 /// assert_eq!((u14::MAX - u14::new(2)).checked_add(u14::new(3)), None);
                 /// ```
@@ -763,7 +767,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u14, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(u14::new(5).checked_mul(u14::new(1)), Some(u14::new(5)));
                 /// assert_eq!(u14::MAX.checked_mul(u14::new(2)), None);
                 /// ```
@@ -919,7 +923,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::u6;
+                /// # use arbitrary_int::prelude::*;
                 /// assert_eq!(u6::new(0b10_1010).reverse_bits(), u6::new(0b01_0101));
                 /// assert_eq!(u6::new(0), u6::new(0).reverse_bits());
                 /// ```
@@ -936,7 +940,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u7, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// let n = u7::new(0b100_1100);
                 /// assert_eq!(n.count_ones(), 3);
                 ///
@@ -959,7 +963,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u7, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// let zero = u7::new(0);
                 /// assert_eq!(zero.count_zeros(), 7);
                 ///
@@ -980,7 +984,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u7, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// let n = !(u7::MAX >> 2);
                 /// assert_eq!(n.leading_ones(), 2);
                 ///
@@ -1002,7 +1006,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u7, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// let n = u7::MAX >> 2;
                 /// assert_eq!(n.leading_zeros(), 2);
                 ///
@@ -1031,7 +1035,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u7, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// let n = u7::new(0b1010111);
                 /// assert_eq!(n.trailing_ones(), 3);
                 ///
@@ -1053,7 +1057,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::{u7, Number};
+                /// # use arbitrary_int::prelude::*;
                 /// let n = u7::new(0b010_1000);
                 /// assert_eq!(n.trailing_zeros(), 3);
                 ///
@@ -1080,7 +1084,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::u6;
+                /// # use arbitrary_int::prelude::*;
                 /// let n = u6::new(0b10_1010);
                 /// let m = u6::new(0b01_0101);
                 ///
@@ -1107,7 +1111,7 @@ macro_rules! uint_impl {
                 /// Basic usage:
                 ///
                 #[doc = concat!(" ```", $doctest_attr)]
-                /// # use arbitrary_int::u6;
+                /// # use arbitrary_int::prelude::*;
                 /// let n = u6::new(0b10_1010);
                 /// let m = u6::new(0b01_0101);
                 ///
@@ -1143,7 +1147,7 @@ uint_impl!(
 // Arithmetic implementations
 impl<T, const BITS: usize> Add for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: PartialEq
         + Copy
         + BitAnd<T, Output = T>
@@ -1168,7 +1172,7 @@ where
 
 impl<T, const BITS: usize> AddAssign for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: PartialEq
         + Eq
         + Not<Output = T>
@@ -1190,7 +1194,7 @@ where
 
 impl<T, const BITS: usize> Sub for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: Copy + BitAnd<T, Output = T> + Sub<T, Output = T>,
 {
     type Output = UInt<T, BITS>;
@@ -1205,7 +1209,7 @@ where
 
 impl<T, const BITS: usize> SubAssign for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: Copy + SubAssign<T> + BitAnd<T, Output = T> + BitAndAssign<T> + Sub<T, Output = T>,
 {
     fn sub_assign(&mut self, rhs: Self) {
@@ -1217,7 +1221,7 @@ where
 
 impl<T, const BITS: usize> Mul for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: PartialEq + Copy + BitAnd<T, Output = T> + Not<Output = T> + Mul<T, Output = T> + From<u8>,
 {
     type Output = UInt<T, BITS>;
@@ -1239,7 +1243,7 @@ where
 
 impl<T, const BITS: usize> MulAssign for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: PartialEq
         + Eq
         + Not<Output = T>
@@ -1261,7 +1265,7 @@ where
 
 impl<T, const BITS: usize> Div for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: PartialEq + Div<T, Output = T>,
 {
     type Output = UInt<T, BITS>;
@@ -1277,7 +1281,7 @@ where
 
 impl<T, const BITS: usize> DivAssign for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: PartialEq + DivAssign<T>,
 {
     fn div_assign(&mut self, rhs: Self) {
@@ -1287,7 +1291,7 @@ where
 
 impl<T, const BITS: usize> BitAnd for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: Copy
         + BitAnd<T, Output = T>
         + Sub<T, Output = T>
@@ -1359,7 +1363,7 @@ where
 
 impl<T, const BITS: usize> Not for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: Copy
         + BitAnd<T, Output = T>
         + BitXor<T, Output = T>
@@ -1379,7 +1383,7 @@ where
 
 impl<T, TSHIFTBITS, const BITS: usize> Shl<TSHIFTBITS> for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: Copy
         + BitAnd<T, Output = T>
         + Shl<TSHIFTBITS, Output = T>
@@ -1407,7 +1411,7 @@ where
 
 impl<T, TSHIFTBITS, const BITS: usize> ShlAssign<TSHIFTBITS> for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: Copy
         + BitAnd<T, Output = T>
         + BitAndAssign<T>
@@ -1540,7 +1544,7 @@ where
 #[cfg(feature = "borsh")]
 impl<T, const BITS: usize> borsh::BorshSerialize for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: borsh::BorshSerialize,
 {
     fn serialize<W: borsh::io::Write>(&self, writer: &mut W) -> borsh::io::Result<()> {
@@ -1555,11 +1559,11 @@ where
 
 #[cfg(feature = "borsh")]
 impl<
-        T: borsh::BorshDeserialize + PartialOrd<<UInt<T, BITS> as Number>::UnderlyingType>,
+        T: borsh::BorshDeserialize + PartialOrd<<UInt<T, BITS> as Integer>::UnderlyingType>,
         const BITS: usize,
     > borsh::BorshDeserialize for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
 {
     fn deserialize_reader<R: borsh::io::Read>(reader: &mut R) -> borsh::io::Result<Self> {
         // Ideally, we'd want a buffer of size `BITS >> 3` or `size_of::<T>`, but that's not possible
@@ -1622,7 +1626,7 @@ where
 #[cfg(feature = "serde")]
 struct InvalidUIntValueError<T>
 where
-    T: Number,
+    T: Integer,
     T::UnderlyingType: Display,
 {
     value: T::UnderlyingType,
@@ -1631,7 +1635,7 @@ where
 #[cfg(feature = "serde")]
 impl<T> Display for InvalidUIntValueError<T>
 where
-    T: Number,
+    T: Integer,
     T::UnderlyingType: Display,
 {
     fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
@@ -1647,7 +1651,7 @@ where
 #[cfg(feature = "serde")]
 impl<'de, T, const BITS: usize> Deserialize<'de> for UInt<T, BITS>
 where
-    Self: Number<UnderlyingType = T>,
+    Self: Integer<UnderlyingType = T>,
     T: Display + PartialOrd + Deserialize<'de>,
 {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
@@ -1665,7 +1669,7 @@ where
 #[cfg(feature = "schemars")]
 impl<T, const BITS: usize> JsonSchema for UInt<T, BITS>
 where
-    Self: Number,
+    Self: Integer,
 {
     fn schema_name() -> String {
         ["uint", &BITS.to_string()].concat()
@@ -1691,7 +1695,7 @@ where
 #[cfg(feature = "step_trait")]
 impl<T, const BITS: usize> Step for UInt<T, BITS>
 where
-    Self: Number<UnderlyingType = T>,
+    Self: Integer<UnderlyingType = T>,
     T: Copy + Step,
 {
     #[inline]
@@ -1721,7 +1725,7 @@ where
 #[cfg(feature = "num-traits")]
 impl<T, const NUM_BITS: usize> num_traits::WrappingAdd for UInt<T, NUM_BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: PartialEq
         + Eq
         + Copy
@@ -1746,7 +1750,7 @@ where
 #[cfg(feature = "num-traits")]
 impl<T, const NUM_BITS: usize> num_traits::WrappingSub for UInt<T, NUM_BITS>
 where
-    Self: Number,
+    Self: Integer,
     T: PartialEq
         + Eq
         + Copy
@@ -1771,7 +1775,7 @@ where
 #[cfg(feature = "num-traits")]
 impl<T, const NUM_BITS: usize> num_traits::bounds::Bounded for UInt<T, NUM_BITS>
 where
-    Self: Number,
+    Self: Integer,
 {
     fn min_value() -> Self {
         Self::MIN